Split ScoutApmServiceProvider into separate Laravel and Lumen specific tests
