+++ known-issues.xml
@@ -528,10 +528,7 @@
       <code>$type</code>
     </UnusedClosureParam>
   </file>
-  <file src="tests/Unit/Laravel/Providers/ScoutApmServiceProviderTest.php">
-    <DeprecatedMethod occurrences="1">
-      <code>setMethods</code>
-    </DeprecatedMethod>
+  <file src="tests/Unit/Laravel/Providers/ScoutApmServiceProviderTestBase.php">
     <InvalidArrayOffset occurrences="1">
       <code>$commands[1]['StartSpan']</code>
     </InvalidArrayOffset>
@@ -542,9 +539,36 @@
       <code>$commands[1]['StartSpan']['operation']</code>
       <code>json_decode(json_encode($metadata), true)['ApplicationEvent']</code>
     </MixedArrayAccess>
+    <PossiblyInvalidArgument occurrences="1">
+      <code>$this-&gt;application</code>
+    </PossiblyInvalidArgument>
+    <UndefinedClass occurrences="6">
+      <code>$events</code>
+      <code>$events</code>
+      <code>$events</code>
+      <code>$events</code>
+      <code>$events</code>
+      <code>$events</code>
+    </UndefinedClass>
     <UnusedClosureParam occurrences="1">
       <code>$whichViews</code>
     </UnusedClosureParam>
+  </file>
+  <file src="tests/Unit/Laravel/Providers/ScoutApmServiceProviderUsingLaravelTest.php">
+    <DeprecatedMethod occurrences="1">
+      <code>setMethods</code>
+    </DeprecatedMethod>
+  </file>
+  <file src="tests/Unit/Laravel/Providers/ScoutApmServiceProviderUsingLumenTest.php">
+    <DeprecatedMethod occurrences="1">
+      <code>setMethods</code>
+    </DeprecatedMethod>
+    <MixedArrayAssignment occurrences="1">
+      <code>$loadedConfigurations['cache']</code>
+    </MixedArrayAssignment>
+    <MixedAssignment occurrences="1">
+      <code>$loadedConfigurations</code>
+    </MixedAssignment>
   </file>
   <file src="tests/Unit/Laravel/View/Engine/ScoutViewEngineDecoratorTest.php">
     <UnusedClosureParam occurrences="4">
